requirements.txt
pyspark>=3.4.0
delta-spark>=2.4.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.24.0
sqlalchemy>=2.0.0
databricks-sql-connector>=2.9.0
azure-identity>=1.15.0
tenacity>=8.2.0
pydantic>=2.0.0
plotly>=5.17.0
kaleido>=0.2.1